refactor: an attempt to delegate functions to different classes for ChannelListViewModel features.
Took 2 hours 21 minutes
