+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/viewmodel/channels/ChannelListViewModel.kt
@@ -23,45 +23,37 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import io.getstream.chat.android.client.ChatClient
-import io.getstream.chat.android.client.api.models.QueryChannelsRequest
-import io.getstream.chat.android.compose.state.QueryConfig
-import io.getstream.chat.android.compose.state.channels.list.ChannelsState
-import io.getstream.chat.android.compose.state.channels.list.ItemState
 import io.getstream.chat.android.compose.state.channels.list.SearchQuery
+import io.getstream.chat.android.compose.viewmodel.channels.usecases.SearchAndQueryChannels
+import io.getstream.chat.android.compose.viewmodel.channels.usecases.SearchAndQueryMessages
 import io.getstream.chat.android.core.utils.Debouncer
 import io.getstream.chat.android.models.Channel
-import io.getstream.chat.android.models.ChannelMute
 import io.getstream.chat.android.models.ConnectionState
 import io.getstream.chat.android.models.FilterObject
 import io.getstream.chat.android.models.Filters
-import io.getstream.chat.android.models.Message
 import io.getstream.chat.android.models.User
 import io.getstream.chat.android.models.querysort.QuerySorter
 import io.getstream.chat.android.state.event.handler.chat.factory.ChatEventHandlerFactory
-import io.getstream.chat.android.state.extensions.globalState
-import io.getstream.chat.android.state.extensions.queryChannelsAsState
-import io.getstream.chat.android.state.plugin.state.querychannels.ChannelsStateData
-import io.getstream.chat.android.state.plugin.state.querychannels.QueryChannelsState
 import io.getstream.chat.android.ui.common.state.channels.actions.Cancel
 import io.getstream.chat.android.ui.common.state.channels.actions.ChannelAction
 import io.getstream.chat.android.uiutils.extension.defaultChannelListFilter
 import io.getstream.log.taggedLogger
 import io.getstream.result.call.toUnitCall
+import kotlinx.coroutines.CoroutineExceptionHandler
+import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.cancelChildren
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.collectLatest
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.mapLatest
 import kotlinx.coroutines.job
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.plus
-import kotlin.coroutines.cancellation.CancellationException
-import kotlin.coroutines.coroutineContext
 
 /**
  * A state store that represents all the information required to query, filter, show and react to
@@ -77,17 +69,17 @@ import kotlin.coroutines.coroutineContext
  * @param chatEventHandlerFactory The instance of [ChatEventHandlerFactory] used to create [ChatEventHandler].
  * @param searchDebounceMs The debounce time for search queries.
  */
-@Suppress("TooManyFunctions")
 public class ChannelListViewModel(
     public val chatClient: ChatClient,
     initialSort: QuerySorter<Channel>,
-    initialFilters: FilterObject?,
+    private val initialFilters: FilterObject?,
     private val channelLimit: Int = DEFAULT_CHANNEL_LIMIT,
     private val memberLimit: Int = DEFAULT_MEMBER_LIMIT,
     private val messageLimit: Int = DEFAULT_MESSAGE_LIMIT,
     private val chatEventHandlerFactory: ChatEventHandlerFactory = ChatEventHandlerFactory(chatClient.clientState),
     searchDebounceMs: Long = SEARCH_DEBOUNCE_MS,
-) : ViewModel() {
+) : ViewModel(),
+    IChannelViewState by ChannelViewStateImpl(chatClient, initialSort, initialFilters) {
 
     private val logger by taggedLogger("Chat:ChannelListVM")
 
@@ -105,36 +97,6 @@ public class ChannelListViewModel(
      * The debouncer used for search operations.
      */
     private val searchDebouncer = Debouncer(searchDebounceMs, searchScope)
-
-    /**
-     * State flow that keeps the value of the current [FilterObject] for channels.
-     */
-    private val filterFlow: MutableStateFlow<FilterObject?> = MutableStateFlow(initialFilters)
-
-    /**
-     * State flow that keeps the value of the current [QuerySorter] for channels.
-     */
-    private val querySortFlow: MutableStateFlow<QuerySorter<Channel>> = MutableStateFlow(initialSort)
-
-    /**
-     * The currently active query configuration, stored in a [MutableStateFlow]. It's created using
-     * the `initialFilters` parameter and the initial sort, but can be changed.
-     */
-    private val queryConfigFlow = filterFlow.filterNotNull().combine(querySortFlow) { filters, sort ->
-        QueryConfig(filters = filters, querySort = sort)
-    }
-
-    /**
-     * The current state of the search input. When changed, it emits a new value in a flow, which
-     * queries and loads new data.
-     */
-    private val searchQuery = MutableStateFlow<SearchQuery>(SearchQuery.Empty)
-
-    /**
-     * The current state of the channels screen. It holds all the information required to render the UI.
-     */
-    public var channelsState: ChannelsState by mutableStateOf(ChannelsState())
-        private set
 
     /**
      * Currently selected channel, if any. Used to show the bottom drawer information when long
@@ -161,11 +123,6 @@ public class ChannelListViewModel(
     public val user: StateFlow<User?> = chatClient.clientState.user
 
     /**
-     * Gives us the information about the list of channels mutes by the current user.
-     */
-    public val channelMutes: StateFlow<List<ChannelMute>> = chatClient.globalState.channelMutes
-
-    /**
      * Builds the default channel filter, which represents "messaging" channels that the current user is a part of.
      */
     private fun buildDefaultFilter(): Flow<FilterObject> {
@@ -182,261 +139,75 @@ public class ChannelListViewModel(
         return channelMutes.value.any { cid == it.channel.cid }
     }
 
-    /**
-     * Current query channels state that contains filter, sort and other states related to channels query.
-     */
-    private var queryChannelsState: StateFlow<QueryChannelsState?> = MutableStateFlow(null)
+    private val searchAndQueryChannels by lazy {
+        SearchAndQueryChannels(
+            channelLimit = channelLimit,
+            messageLimit = messageLimit,
+            memberLimit = memberLimit,
+            chatClient = chatClient,
+            chatEventHandlerFactory = chatEventHandlerFactory,
+            channelState = this,
+            logger = logger,
+        )
+    }
 
-    /**
-     * The current state of the search Messages. When changed, it emits a new value in a flow, which
-     * queries and loads new data.
-     */
-    private val searchMessageState: MutableStateFlow<SearchMessageState?> = MutableStateFlow(null)
-
-    private var lastNextQuery: QueryChannelsRequest? = null
+    private val searchAndQueryMessages by lazy {
+        SearchAndQueryMessages(chatClient, logger, channelLimit, this)
+    }
 
     /**
      * Combines the latest search query and filter to fetch channels and emit them to the UI.
      */
     init {
+        setupFilters()
+
+        setupSearchAndQuery()
+    }
+
+    private fun setupFilters() {
         if (initialFilters == null) {
             viewModelScope.launch {
                 val filter = buildDefaultFilter().first()
-
-                this@ChannelListViewModel.filterFlow.value = filter
+                filterFlow.value = filter
             }
-        }
-
-        viewModelScope.launch {
-            init()
         }
     }
 
     /**
      * Makes the initial query to request channels and starts observing state changes.
      */
-    private suspend fun init() {
+    @OptIn(ExperimentalCoroutinesApi::class)
+    private fun setupSearchAndQuery() {
         logger.d { "[init] no args" }
         searchQuery.combine(queryConfigFlow) { query, config -> query to config }
-            .collectLatest { (query, config) ->
+            .mapLatest { (query, config) ->
                 when (query) {
                     is SearchQuery.Empty,
                     is SearchQuery.Channels,
                     -> {
-                        searchScope.coroutineContext.cancelChildren()
-                        observeQueryChannels(query.getConfig(config))
+                        searchAndQueryChannels(
+                            searchScope = searchScope,
+                            chListScope = chListScope,
+                            config = query.getConfig(config),
+                        )
                     }
+
                     is SearchQuery.Messages -> {
-                        chListScope.coroutineContext.cancelChildren()
-                        handleSearchQuery(query.query)
-                        observeSearchMessages(query.query)
+                        searchAndQueryMessages(
+                            coroutineScope = chListScope,
+                            query = query.query,
+                            searchDebouncer = searchDebouncer,
+                        )
                     }
                 }
-            }
+            }.catch {
+                logger.e(it) {
+                    "setupSearchAndQuery failed"
+                }
+            }.launchIn(viewModelScope)
     }
 
-    private fun SearchQuery.getConfig(defaultConfig: QueryConfig<Channel>): QueryConfig<Channel> {
-        return when (this) {
-            is SearchQuery.Channels -> defaultConfig.copy(
-                filters = createQueryChannelsFilter(defaultConfig.filters, query),
-            )
-            is SearchQuery.Messages -> defaultConfig
-            is SearchQuery.Empty -> defaultConfig
-        }
-    }
 
-    private suspend fun observeSearchMessages(query: String) = runCatching {
-        logger.d { "[observeSearchMessages] query: '$query'" }
-        searchMessageState.filterNotNull().collectLatest {
-            logger.v { "[observeSearchMessages] state: ${it.stringify()}" }
-            channelsState = channelsState.copy(
-                searchQuery = searchQuery.value,
-                isLoading = it.isLoading,
-                isLoadingMore = it.isLoadingMore,
-                endOfChannels = !it.canLoadMore,
-                channelItems = it.messages.map(ItemState::SearchResultItemState),
-            )
-        }
-    }.onFailure {
-        when (it is CancellationException) {
-            true -> logger.v { "[observeSearchMessages] cancelled('$query')" }
-            else -> logger.e { "[observeSearchMessages] failed: $it" }
-        }
-    }
-
-    private fun handleSearchQuery(query: String) {
-        logger.d { "[handleSearchQuery] query: '$query'" }
-        searchDebouncer.submitSuspendable {
-            searchMessagesForQuery(query)
-        }
-    }
-
-    private suspend fun searchMessagesForQuery(query: String) {
-        logger.d { "[searchMessagesForQuery] query: '$query'" }
-        val channelFilter = filterFlow.value ?: Filters.defaultChannelListFilter(user.value) ?: run {
-            logger.v { "[searchMessagesForQuery] rejected (no channel filter)" }
-            return
-        }
-        val newState = SearchMessageState(query = query, isLoading = true)
-        searchMessageState.value = newState
-        searchMessageState.value = searchMessages(src = "new", newState, channelFilter).also {
-            logger.v { "[searchMessagesForQuery] completed('$query'): ${it.messages.size}" }
-        }
-    }
-
-    private suspend fun loadMoreQueryMessages() {
-        logger.d { "[loadMoreQueryMessages] no args" }
-        val channelFilter = filterFlow.value ?: Filters.defaultChannelListFilter(user.value) ?: run {
-            logger.v { "[loadMoreQueryMessages] rejected (no channel filter)" }
-            return
-        }
-        val currentState = searchMessageState.value ?: run {
-            logger.v { "[loadMoreQueryMessages] rejected (no current state)" }
-            return
-        }
-        if (currentState.isLoading) {
-            logger.v { "[loadMoreQueryMessages] rejected (already loading)" }
-            return
-        }
-        if (currentState.isLoadingMore) {
-            logger.v { "[loadMoreQueryMessages] rejected (already loading more)" }
-            return
-        }
-        if (!currentState.canLoadMore) {
-            logger.v { "[loadMoreQueryMessages] rejected (end of messages)" }
-            return
-        }
-        val query = currentState.query
-        logger.v { "[loadMoreQueryMessages] query: 'query'" }
-        val newState = currentState.copy(isLoadingMore = true)
-        searchMessageState.value = newState
-        searchMessageState.value = searchMessages(src = "more", newState, channelFilter).also {
-            logger.v { "[loadMoreQueryMessages] completed('$query'): ${it.messages.size}" }
-        }
-    }
-
-    /**
-     * Searches for messages based on the current query.
-     */
-    private suspend fun searchMessages(
-        src: String,
-        currentState: SearchMessageState,
-        channelFilter: FilterObject,
-    ): SearchMessageState {
-        val offset = currentState.messages.size
-        val limit = channelLimit
-        logger.v { "[searchMessages] #$src; query: '${currentState.query}', offset: $offset, limit: $limit" }
-        val result = chatClient.searchMessages(
-            channelFilter = channelFilter,
-            messageFilter = Filters.autocomplete("text", currentState.query),
-            offset = offset,
-            limit = limit,
-        ).await()
-        return when (result) {
-            is io.getstream.result.Result.Success -> {
-                logger.v { "[searchMessages] #$src; completed(messages.size: ${result.value.messages.size})" }
-                currentState.copy(
-                    messages = currentState.messages + result.value.messages,
-                    isLoading = false,
-                    isLoadingMore = false,
-                    canLoadMore = result.value.messages.size >= limit,
-                )
-            }
-            is io.getstream.result.Result.Failure -> {
-                logger.e { "[searchMessages] #$src; failed: ${result.value}" }
-                currentState.copy(
-                    isLoading = false,
-                    isLoadingMore = false,
-                    canLoadMore = true,
-                )
-            }
-        }
-    }
-
-    private suspend fun observeQueryChannels(config: QueryConfig<Channel>) = runCatching {
-        val queryChannelsRequest = QueryChannelsRequest(
-            filter = config.filters,
-            querySort = config.querySort,
-            limit = channelLimit,
-            messageLimit = messageLimit,
-            memberLimit = memberLimit,
-        )
-
-        logger.d { "[observeQueryChannels] request: $queryChannelsRequest" }
-        queryChannelsState = chatClient.queryChannelsAsState(
-            request = queryChannelsRequest,
-            chatEventHandlerFactory = chatEventHandlerFactory,
-            coroutineScope = chListScope,
-        )
-
-        queryChannelsState.filterNotNull().collectLatest { queryChannelsState ->
-            channelMutes.combine(queryChannelsState.channelsStateData, ::Pair)
-                .map { (channelMutes, state) ->
-                    when (state) {
-                        ChannelsStateData.NoQueryActive,
-                        ChannelsStateData.Loading,
-                        -> channelsState.copy(
-                            isLoading = true,
-                            searchQuery = searchQuery.value,
-                        ).also {
-                            logger.d { "[observeQueryChannels] state: Loading" }
-                        }
-                        ChannelsStateData.OfflineNoResults -> {
-                            logger.v { "[observeQueryChannels] state: OfflineNoResults(channels are empty)" }
-                            channelsState.copy(
-                                isLoading = false,
-                                channelItems = emptyList(),
-                                searchQuery = searchQuery.value,
-                            )
-                        }
-                        is ChannelsStateData.Result -> {
-                            logger.v { "[observeQueryChannels] state: Result(channels.size: ${state.channels.size})" }
-                            channelsState.copy(
-                                isLoading = false,
-                                channelItems = createChannelItems(state.channels, channelMutes),
-                                isLoadingMore = false,
-                                endOfChannels = queryChannelsState.endOfChannels.value,
-                                searchQuery = searchQuery.value,
-                            )
-                        }
-                    }
-                }.collectLatest { newState -> channelsState = newState }
-        }
-    }.onFailure {
-        when (it is CancellationException) {
-            true -> logger.v { "[observeQueryChannels] cancelled" }
-            else -> logger.e { "[observeQueryChannels] failed: $it" }
-        }
-    }
-
-    /**
-     * Creates a filter that is used to query channels.
-     *
-     * If the [searchQuery] is empty, then returns the original [filter] provided by the user.
-     * Otherwise, returns a wrapped [filter] that also checks that the channel name match the
-     * [searchQuery].
-     *
-     * @param filter The filter that was passed by the user.
-     * @param searchQuery The search query used to filter the channels.
-     *
-     * @return The filter that will be used to query channels.
-     */
-    private fun createQueryChannelsFilter(filter: FilterObject, searchQuery: String): FilterObject {
-        return if (searchQuery.isNotEmpty()) {
-            Filters.and(
-                filter,
-                Filters.or(
-                    Filters.and(
-                        Filters.autocomplete("member.user.name", searchQuery),
-                        Filters.notExists("name"),
-                    ),
-                    Filters.autocomplete("name", searchQuery),
-                ),
-            )
-        } else {
-            filter
-        }
-    }
 
     /**
      * Changes the currently selected channel state. This updates the UI state and allows us to observe
@@ -502,50 +273,19 @@ public class ChannelListViewModel(
         when (searchQuery.value) {
             is SearchQuery.Empty,
             is SearchQuery.Channels,
-            -> chListScope.launch { loadMoreQueryChannels() }
-            is SearchQuery.Messages,
-            -> searchScope.launch { loadMoreQueryMessages() }
-        }
-    }
+            -> chListScope.launch(CoroutineExceptionHandler { _, throwable ->
+                logger.e(throwable) {
+                    "failed to loadMoreQueryChannels"
+                }
+            }) { searchAndQueryChannels.loadMoreQueryChannels() }
 
-    private suspend fun loadMoreQueryChannels() {
-        logger.d { "[loadMoreQueryChannels] no args" }
-        val currentFilter = filterFlow.value
-        if (currentFilter == null) {
-            logger.v { "[loadMoreQueryChannels] rejected (no current filter)" }
-            return
+            is SearchQuery.Messages,
+            -> searchScope.launch(CoroutineExceptionHandler { _, throwable ->
+                logger.e(throwable) {
+                    "failed to loadMoreQueryMessages"
+                }
+            }) { searchAndQueryMessages.loadMoreQueryMessages() }
         }
-        val currentQuery = queryChannelsState.value?.nextPageRequest?.value
-        if (currentQuery == null) {
-            logger.v { "[loadMoreQueryChannels] rejected (no current query)" }
-            return
-        }
-        if (channelsState.endOfChannels) {
-            logger.v { "[loadMoreQueryChannels] rejected (end of channels)" }
-            return
-        }
-        if (channelsState.isLoadingMore) {
-            logger.v { "[loadMoreQueryChannels] rejected (already loading more)" }
-            return
-        }
-        val nextQuery = currentQuery.copy(
-            filter = createQueryChannelsFilter(currentFilter, searchQuery.value.query),
-            querySort = querySortFlow.value,
-        )
-        if (lastNextQuery == nextQuery) {
-            logger.v { "[loadMoreQueryChannels] rejected (same query)" }
-            return
-        }
-        lastNextQuery = nextQuery
-        logger.v { "[loadMoreQueryChannels] offset: ${nextQuery.offset}, limit: ${nextQuery.limit}" }
-        channelsState = channelsState.copy(isLoadingMore = true)
-        val result = chatClient.queryChannels(nextQuery).await()
-        if (result.isSuccess) {
-            logger.v { "[loadMoreQueryChannels] completed; channels.size: ${result.getOrNull()?.size}" }
-        } else {
-            logger.e { "[loadMoreQueryChannels] failed: ${result.errorOrNull()}" }
-        }
-        channelsState = channelsState.copy(isLoadingMore = false)
     }
 
     /**
@@ -624,21 +364,6 @@ public class ChannelListViewModel(
         selectedChannel.value = null
     }
 
-    /**
-     * Creates a list of [ChannelItemState] that represents channel items we show in the list of channels.
-     *
-     * @param channels The channels to show.
-     * @param channelMutes The list of channels muted for the current user.
-     *
-     */
-    private fun createChannelItems(
-        channels: List<Channel>,
-        channelMutes: List<ChannelMute>,
-    ): List<ItemState.ChannelItemState> {
-        val mutedChannelIds = channelMutes.map { channelMute -> channelMute.channel.cid }.toSet()
-        return channels.map { ItemState.ChannelItemState(it, it.cid in mutedChannelIds) }
-    }
-
     internal companion object {
         /**
          * Default value of number of channels to return when querying channels.
@@ -659,23 +384,5 @@ public class ChannelListViewModel(
          * Debounce time for search queries.
          */
         private const val SEARCH_DEBOUNCE_MS = 200L
-    }
-
-    private data class SearchMessageState(
-        val query: String = "",
-        val canLoadMore: Boolean = true,
-        val messages: List<Message> = emptyList(),
-        val isLoading: Boolean = false,
-        val isLoadingMore: Boolean = false,
-    ) {
-
-        fun stringify(): String {
-            return "SearchMessageState(" +
-                "query='$query', " +
-                "messages.size=${messages.size}, " +
-                "isLoading=$isLoading, " +
-                "isLoadingMore=$isLoadingMore, " +
-                "canLoadMore=$canLoadMore)"
-        }
     }
 }
